+++ internal/tests/cluster/ipv6_listener_test.go
@@ -60,7 +60,7 @@ func TestIPv6Listener(t *testing.T) {
 				// expecting it we'll see an out-of-date entry
 				return true
 			}
-			assert.WithinDuration(time.Now(), v.(time.Time), 35*time.Second)
+			assert.WithinDuration(time.Now(), v.(time.Time), 60*time.Second)
 			delete(workerMap, k.(string))
 			return true
 		})